requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
ollama
pytz
requests
pyttsx3
SpeechRecognition
PyAudio

# Dependencies from services/requirements.txt for web scraping/automation
beautifulsoup4
pytesseract
Pillow
pydantic
lxml
#undetected-chromedriver
#selenium